Fix the order of the arguments in res_mkquery()
As noted by camm73 in his pull request
https://github.com/funnydog/ds-nslookup/pull/1, the order of class and
type arguments were inverted.

By pure luck the code worked as expected because the actual value of
the constants matched the correct values for the query.

Let's use the right constants for the class and type fields.

Kudos to camm73 who spotted the bug.